chore: 컴포넌트 구조 수정 - index 폴더 추가
