+++ conponents/(tabs)/index/FooterSection.tsx
