+++ conponents/(tabs)/index/MyStatusSection/AnimatedImages/HomeImage.tsx
