+++ conponents/(tabs)/index/MyStatusSection/LinearBackground/HomeLinearBackground.tsx
