+++ conponents/(tabs)/index/MyStatusSection/TravelCard/TravelCard.tsx
