+++ conponents/(tabs)/index/NearActivitySection/FilterModal.tsx
