+++ conponents/(tabs)/index/NearActivitySection/index.tsx
